TQMa28: Remove cmd_spi and spi pinmux
The spi command issues full duplex transfers by providing pointers to a data in and a data out buffer when a spi_xfer is initiated.
The mxs_spi driver prohibits transfers with both pointers being set as the mx28 hardware is only capable of half-duplex transfers.
As long as there is no support for half-duplex transfers in the spi command this function is not usable.

Signed-off-by: Michael Krummsdorf <michael.krummsdorf@tqs.de>