TQMa28 - Fix install script calculation errors
The script had issues when comparing hexdecimals numbers.

The setexpr command works at base 16 but returns no 0x prefix.
The term "if test 1a0 -gt 2" (where 1a0 is a setexpr returned number)
returnes false as the command 'test' misinterpretes the 1a0 as decimal
number thus interpreting its value as 1, which obviously is not >= 2.

Using the 'itest' command instead of 'test' solves the issue.

Signed-off-by: Michael Krummsdorf <michael.krummsdorf@tqs.de>